Fix VersionUtil::getVersion()
Fix VersionUtil::getVersion(). The new way the version is loaded is
inspired by Debezium connectors and Scylla CDC Source Connector.

The version is stored in "version.properties" resource file and Maven
on build-time fills this file with a version number. VersionUtil
can then read it in runtime, by loading the properties file.

Fixes #39.