Make commit after every batch of inserts optional, with the default not to commit.
For at least once semantics, it is not necessary to commit after every call to put(), as Kafka
Connect will commit the kafka Offsets periodically. Doing it periodically will lead to less
writes to the __consumer_offsets topic, and hence, less load on the Kafka brokers.

The default is false, i.e not to commit offset after every set of INSERT operations. This
is a change in behaviour of Kafka Connect ScyllaDB, but it leads to a significant decrease
in number of produce requests on the Kafka brokers, which in production environments
with high load is quite visible in Broker CPU consumption metrics.